# docs/openapi.yaml
openapi: '3.1.0'
info:
  title: AL-SAQI Internal Audit System API
  version: '1.0.0'
  description: >
    REST API for the AL-SAQI internal audit management system.
    Handles audit planning, findings, recommendations, risk management,
    correspondence, compliance, and user management.
servers:
  - url: /api
    description: Main API server
security:
  - BearerAuth: []
    CsrfToken: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 JWT token passed via httpOnly cookie or Authorization header
    CsrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: CSRF token required for state-changing requests (POST, PUT, PATCH, DELETE)
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Admin, Internal Auditor, Compliance Officer, Risk Officer, Manager, Viewer]
        department:
          type: string
          nullable: true
        job_title_id:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        status:
          type: string
          enum: [Active, Suspended, Archived]
        phone_number:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
paths:
...
  /auth/login:
    post:
      summary: User login
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [usernameOrEmail, password]
              properties:
                usernameOrEmail:
                  type: string
                password:
                  type: string
                rememberMe:
                  type: boolean
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
        '401':
          description: Invalid credentials
...